change : int to size_t

+++ src/bitmap/bitmap.cpp
@@ -121,7 +121,7 @@ void Bitmap::Prepare( const wxImage& image, int pos, struct SBitmap& bmp )
     bmp.SetBitmap( wxBitmap(image) );
 }
 
-void Bitmap::Add( wxImage& image, int idx )
+void Bitmap::Add( wxImage& image, size_t idx )
 {
     Size::Prepare( image, m_flagSize, m_parent, m_scaleParent );
     struct SBitmap& bmp = m_item.at(idx);
